Decouple diffdisk (misnomer) from basedisk
Fix issue 1706.

Now basedisk is immediately renamed/converted to diffdisk
(not a diff despite the name) by the VM driver,
except when basedisk is an ISO9660 image.

Decoupling diffdisk from basedisk will eliminate the overhead of differencing I/O
and save some disk space.

I cannot remember why I designed the disk to be split into basedisk and diffdisk.
Maybe it was to allow `limactl factory-reset` to retain the initial state,
although the command was apparently never implemented in that way.
Maybe it was to allow multiple instances to share the same basedisk,
although it was never implemented in that way, either.

+++ pkg/driverutil/disk.go
@@ -18,6 +18,8 @@ import (
 	"github.com/lima-vm/lima/v2/pkg/limatype/filenames"
 )
 
+// EnsureDiskRaw just converts baseDisk (can be qcow2) to diffDisk (raw), unless baseDisk is an ISO9660 image.
+// Note that "diffDisk" is a misnomer, it is actually created as a full disk since Lima v2.1.
 func EnsureDiskRaw(ctx context.Context, inst *limatype.Instance) error {
 	diffDisk := filepath.Join(inst.Dir, filenames.DiffDisk)
 	if _, err := os.Stat(diffDisk); err == nil || !errors.Is(err, os.ErrNotExist) {
@@ -33,26 +35,33 @@ func EnsureDiskRaw(ctx context.Context, inst *limatype.Instance) error {
 	if diskSize == 0 {
 		return nil
 	}
-	isBaseDiskISO, err := iso9660util.IsISO9660(baseDisk)
-	if err != nil {
-		return err
-	}
-	if isBaseDiskISO {
-		// Create an empty data volume (sparse)
-		diffDiskF, err := os.Create(diffDisk)
+	var isBaseDiskISO bool
+	if _, err := os.Stat(baseDisk); !errors.Is(err, os.ErrNotExist) {
+		isBaseDiskISO, err = iso9660util.IsISO9660(baseDisk)
 		if err != nil {
 			return err
 		}
+		if isBaseDiskISO {
+			// Create an empty data volume (sparse)
+			diffDiskF, err := os.Create(diffDisk)
+			if err != nil {
+				return err
+			}
 
-		err = diskUtil.MakeSparse(ctx, diffDiskF, 0)
-		if err != nil {
-			diffDiskF.Close()
-			return fmt.Errorf("failed to create sparse diff disk %q: %w", diffDisk, err)
+			err = diskUtil.MakeSparse(ctx, diffDiskF, 0)
+			if err != nil {
+				diffDiskF.Close()
+				return fmt.Errorf("failed to create sparse diff disk %q: %w", diffDisk, err)
+			}
+			return diffDiskF.Close()
 		}
-		return diffDiskF.Close()
 	}
-	if err = diskUtil.ConvertToRaw(ctx, baseDisk, diffDisk, &diskSize, false); err != nil {
+	// "diffdisk" is a misnomer, it is actually created as a full disk since Lima v2.1.
+	if err := diskUtil.ConvertToRaw(ctx, baseDisk, diffDisk, &diskSize, false); err != nil {
 		return fmt.Errorf("failed to convert %q to a raw disk %q: %w", baseDisk, diffDisk, err)
 	}
-	return err
+	if err := os.RemoveAll(baseDisk); err != nil {
+		return err
+	}
+	return nil
 }
